requirements.txt
websockets>=12.0
speechrecognition>=3.10
psutil>=5.9
pyautogui>=0.9
requests>=2.31
pywebview>=5.0
groq>=0.9.0
openai-whisper>=20231117
httpx>=0.27.0
